Move the metrics server into a runner package
We'll be adding additional runners to this package
